requirements-dev.txt
# Development dependencies for Menteebot CI/CD Demo
# Include production dependencies
-r requirements.txt

# Testing framework
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-flask==1.3.0

# Code quality and linting
flake8==6.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Security scanning
bandit==1.7.5
safety==2.3.5

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Development tools
pre-commit==3.6.0
tox==4.11.4

# Debugging
ipdb==0.13.13
ipython==8.18.1

# API testing
pytest-asyncio==0.21.1
httpx==0.25.2

# Performance testing
locust==2.17.0

# Docker testing
testcontainers==3.7.1 